Reorganise into entity logic and API interactions with hub
